♻️ refactor :: 기존 댓글 타입을 관리하던 modal 리팩토링
